Move base class to deep/agents/
