+++ genrl/deep/agents/__init__.py
@@ -1,4 +1,5 @@
 from genrl.deep.agents.a2c import A2C  # noqa
+from genrl.deep.agents.base import BaseAgent, OnPolicyAgent  # noqa
 from genrl.deep.agents.ddpg import DDPG  # noqa
 from genrl.deep.agents.dqn import DQN  # noqa
 from genrl.deep.agents.ppo1 import PPO1  # noqa
