+++ genrl/deep/agents/a2c/a2c.py
@@ -8,7 +8,8 @@ import torch.optim as opt
 from torch.autograd import Variable
 
 from ....environments.vec_env import VecEnv
-from ...common import OnPolicyAgent, RolloutBuffer, get_env_properties, get_model
+from ...common import RolloutBuffer, get_env_properties, get_model
+from ..base import OnPolicyAgent
 
 
 class A2C(OnPolicyAgent):
